// typechain-types/NodeOperatorFactory.ts
export interface NodeOperatorFactoryInterface extends ethers.utils.Interface {
  functions: {
    "create(string,address)": FunctionFragment;
    "instances(uint256)": FunctionFragment;
    "namespace(string)": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "create",
    values: [string, string]
  ): string;
  encodeFunctionData(
    functionFragment: "instances",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(functionFragment: "namespace", values: [string]): string;

  decodeFunctionResult(functionFragment: "create", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "instances", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "namespace", data: BytesLike): Result;

  events: {
    "NodeOperatorCreated(address,address,address,uint256)": EventFragment;
  };

  getEvent(nameOrSignatureOrTopic: "NodeOperatorCreated"): EventFragment;
}

export type NodeOperatorCreatedEvent = TypedEvent<
  [string, string, string, BigNumber],
  {
    creator: string;
    nodeOperator: string;
    stablecoin: string;
    timestamp: BigNumber;
  }
>;

export type NodeOperatorCreatedEventFilter =
  TypedEventFilter<NodeOperatorCreatedEvent>;

export interface NodeOperatorFactory extends BaseContract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: NodeOperatorFactoryInterface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    create(
      name: string,
      stablecoin: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<ContractTransaction>;

    instances(arg0: BigNumberish, overrides?: CallOverrides): Promise<[string]>;

    namespace(arg0: string, overrides?: CallOverrides): Promise<[string]>;
  };

  create(
    name: string,
    stablecoin: string,
    overrides?: Overrides & { from?: string | Promise<string> }
  ): Promise<ContractTransaction>;

  instances(arg0: BigNumberish, overrides?: CallOverrides): Promise<string>;

  namespace(arg0: string, overrides?: CallOverrides): Promise<string>;

  callStatic: {
    create(
      name: string,
      stablecoin: string,
      overrides?: CallOverrides
    ): Promise<string>;

    instances(arg0: BigNumberish, overrides?: CallOverrides): Promise<string>;

    namespace(arg0: string, overrides?: CallOverrides): Promise<string>;
  };

  filters: {
    "NodeOperatorCreated(address,address,address,uint256)"(
      creator?: string | null,
      nodeOperator?: null,
      stablecoin?: null,
      timestamp?: null
    ): NodeOperatorCreatedEventFilter;
    NodeOperatorCreated(
      creator?: string | null,
      nodeOperator?: null,
      stablecoin?: null,
      timestamp?: null
    ): NodeOperatorCreatedEventFilter;
  };

  estimateGas: {
    create(
      name: string,
      stablecoin: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<BigNumber>;

    instances(
      arg0: BigNumberish,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    namespace(arg0: string, overrides?: CallOverrides): Promise<BigNumber>;
  };

  populateTransaction: {
    create(
      name: string,
      stablecoin: string,
      overrides?: Overrides & { from?: string | Promise<string> }
    ): Promise<PopulatedTransaction>;

    instances(
      arg0: BigNumberish,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    namespace(
      arg0: string,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;
  };
}